+++ twitter-kafka-producer/src/main/java/com/github/carlosasrc/samplekafkaelasticsearch/twitterkafkaproducer/TwitterConfig.java
@@ -1,4 +1,4 @@
-package com.github.carlosasrc.samplekafkaelasticsearch.twittermicroservice;
+package com.github.carlosasrc.samplekafkaelasticsearch.twitterkafkaproducer;
 
 import com.google.common.collect.Lists;
 import com.twitter.hbc.ClientBuilder;
