+++ twitter-kafka-producer/src/main/java/com/github/carlosasrc/samplekafkaelasticsearch/twitterkafkaproducer/TwitterKafkaProducerApplication.java
@@ -1,0 +1,11 @@
+package com.github.carlosasrc.samplekafkaelasticsearch.twitterkafkaproducer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TwitterKafkaProducerApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(TwitterKafkaProducerApplication.class, args);
+    }
+}
